divide calcAbility
